requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-telegram-bot>=20.0

# Trading and financial data
alpha-vantage>=2.3.1

# Data processing and analysis
plotly>=5.15.0
matplotlib>=3.7.0
scipy>=1.11.0

# Database
sqlalchemy>=2.0.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
playwright>=1.37.0
coverage>=7.3.0

# Code quality
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Async support
aiohttp>=3.8.0